+++ sickrage/core/webserver/__init__.py
@@ -21,6 +21,7 @@ from __future__ import unicode_literals
 import os
 import shutil
 import socket
+import threading
 
 from tornado.httpserver import HTTPServer
 from tornado.web import Application, RedirectHandler, StaticFileHandler
@@ -45,11 +46,9 @@ class StaticImageHandler(StaticFileHandler):
         return super(StaticImageHandler, self).get(path, include_body)
 
 
-class srWebServer(object):
+class srWebServer(threading.Thread):
     def __init__(self):
-        super(srWebServer, self).__init__()
-        self.name = "TORNADO"
-        self.daemon = True
+        super(srWebServer, self).__init__(name="TORNADO")
         self.started = False
         self.video_root = None
         self.api_root = None